fix requirejs error
