+++ src/core/algorithms.js
@@ -5,9 +5,10 @@ Currently used only for adding 'assert' class to algorithm lists
 
 export const name = "core/algorithms";
 
+/** @return {Promise<any>} */
 async function loadStyle() {
   try {
-    return (await import("text!../../assets/algorithms.css")).default;
+    return await import("text!../../assets/algorithms.css");
   } catch {
     const res = await fetch(
       new URL("../../assets/algorithms.css", import.meta.url)
