+++ src/core/best-practices.js
@@ -9,9 +9,10 @@ import { pub } from "./pubsubhub";
 
 export const name = "core/best-practices";
 
+/** @return {Promise<any>} */
 async function loadStyle() {
   try {
-    return (await import("text!../../assets/bp.css")).default;
+    return await import("text!../../assets/bp.css");
   } catch {
     const res = await fetch(new URL("../../assets/bp.css", import.meta.url));
     return await res.text();
