+++ src/core/caniuse.js
@@ -1,3 +1,4 @@
+// @ts-check
 /**
  * Module: "core/caniuse"
  * Adds a caniuse support table for a "feature" #1238
@@ -41,9 +42,10 @@ const supportTitles = new Map([
   ["d", "Disabled by default (needs to enabled)."],
 ]);
 
+/** @return {Promise<any>} */
 async function loadStyle() {
   try {
-    return (await import("text!../../assets/caniuse.css")).default;
+    return await import("text!../../assets/caniuse.css");
   } catch {
     const res = await fetch(
       new URL("../../assets/caniuse.css", import.meta.url)
@@ -90,13 +92,13 @@ export async function run(conf) {
     }
     resolve(content);
   });
-  const definitionPair = hyperHTML.bind(document.createDocumentFragment())`
+  const definitionPair = hyperHTML`
     <dt class="caniuse-title">Can I Use this API?</dt>
     <dd class="caniuse-stats">${{
       any: contentPromise,
       placeholder: "Fetching data from caniuse.com...",
     }}</dd>`;
-  headDlElem.appendChild(definitionPair);
+  headDlElem.append(...definitionPair.childNodes);
   await contentPromise;
 
   // remove from export
@@ -168,7 +170,7 @@ function createTableHTML(featureURL, stats) {
 
 /**
  * Add a browser and it's support to table.
- * @param {[ string, ApiResponse["browserName"] ]}
+ * @param {[ string, ApiResponse["browserName"] ]} args
  */
 function addBrowser([browserName, browserData]) {
   /** @param {string[]} supportKeys */
@@ -182,7 +184,7 @@ function addBrowser([browserName, browserData]) {
     };
   };
 
-  /** @param {[string, string[]]} */
+  /** @param {[string, string[]]} args */
   const addLatestVersion = ([version, supportKeys]) => {
     const { className, title } = getSupport(supportKeys);
     const buttonText = `${BROWSERS.get(browserName) || browserName} ${version}`;
@@ -190,7 +192,7 @@ function addBrowser([browserName, browserData]) {
       <button class="${className}" title="${title}">${buttonText}</button>`;
   };
 
-  /** @param {[string, string[]]} */
+  /** @param {[string, string[]]} args */
   const addBrowserVersion = ([version, supportKeys]) => {
     const { className, title } = getSupport(supportKeys);
     return `<li class="${className}" title="${title}">${version}</li>`;
