+++ src/core/examples.js
@@ -57,9 +57,10 @@ function makeTitle(conf, elem, num, report) {
   `;
 }
 
+/** @return {Promise<any>} */
 async function loadStyle() {
   try {
-    return (await import("text!../../assets/examples.css")).default;
+    return await import("text!../../assets/examples.css");
   } catch {
     const res = await fetch(
       new URL("../../assets/examples.css", import.meta.url)
