+++ src/core/ui.js
@@ -18,9 +18,10 @@ export const name = "core/ui";
 // Opportunistically inserts the style, with the chance to reduce some FOUC
 insertStyle();
 
+/** @return {Promise<any>} */
 async function loadStyle() {
   try {
-    return (await import("text!../../assets/ui.css")).default;
+    return await import("text!../../assets/ui.css");
   } catch {
     const res = await fetch(new URL("../../assets/ui.css", import.meta.url));
     return await res.text();
