+++ src/core/worker.js
@@ -18,9 +18,10 @@ const hint = {
 const link = createResourceHint(hint);
 document.head.appendChild(link);
 
+/** @return {Promise<any>} */
 async function loadWorkerScript() {
   try {
-    return (await import("text!../../worker/respec-worker.js")).default;
+    return await import("text!../../worker/respec-worker.js");
   } catch {
     const res = await fetch(
       new URL("../../worker/respec-worker.js", import.meta.url)
